move all pouchdb related utils from the root to version-1.2
