+++ pkg/db/src/implementations/version-1.2/stock.ts
@@ -4,8 +4,7 @@ import { debug, StockMap, wrapIter } from "@librocco/shared";
 
 import { InventoryDatabaseInterface, WarehouseData, NoteData } from "./types";
 
-import { newChangesStream } from "@/utils/pouchdb";
-import { versionId } from "./utils";
+import { newChangesStream, versionId } from "./utils";
 import { isBookRow } from "@/utils/misc";
 
 export interface StockInterface {
